Implement faction role commands
These commands manage so-called faction roles, which are vanity roles
where only one can ever be had by a user at a time so that their color
is set to that of the role regardless of any other roles they have. They
should be under the Bot role but listed higher than all normal vanity
roles so that they override those color-wise.

The commands are !listfactions, !addfaction, and !removefaction. The
commands are intelligent about ensuring that any roles operated on are
actually faction roles; they check that the faction role ends in '
Faction' and has a corresponding role with the same name as the faction
role but without this suffix. The !addfaction command will remove any
existing faction before giving the user this faction, thus preventing
users from having multiple factions. This command also allows omitting
the ' Faction' suffix when specifying the desired role and ignores case.
The !removefaction command will remove any faction roles the user has.