Use a dedicated Bot Faction role for factions
For factions, look for a 'Bot Faction' role as the one below which we
allow roles to be considered as factions. This way servers can have
roles that are listed above the 'Bot' role used for normal vanity roles
so that they aren't modifiable via !addrole/!removerole and where such
roles don't override the color of faction roles. Faction roles can be
listed above these roles so that they set the user's color properly.
Prior to this change, to make a role unassignable as a vanity role, it
had to be listed above the faction roles, hence it also had to override
faction colors.

The 'Bot Faction' role should be assigned to the bot along with the
'Bot' role, but must be higher than this role, and likewise all faction
roles must be above the 'Bot' role but below 'Bot Faction'.