Disable Discord previews for URLS
Cerebot outputs a lot of URLs, and the previews for these tend to be
annoying, especially when an image is involved. For certain types of
queries like %git, we'd ideally do some kind of custom output that
employs html markup. I'd need to look more into the Discord side of this
to see what could feasibly be done. Until then, use angle brackets to
disable the preview on URLs.