Updates to DB and bot command support from beem
This commit adapts Cerebot to changes to database and bot command
handling introduced in gammafunk/beem@fd152b13 and
gammafunk/beem@63512f34. It adds DB support to Cerebot and uses it to
implement a server allow list, per-server permissions based on moderator
roles, and DCSS nick mapping.

We now have enums to describe permission levels are used, and bot admin
status is stored directly in the database. The bot owner is
automatically given bot admin permission in the database after login.
The server admin and server moderator (see below) permissions levels are
determined on the fly using the Discord API.  The !allowserver and
!disallowserver commands can be used to allow the bot to respond (or
not) to servers that have added it on the fly, making management of it
as a public bot much easier.

The !dcssnick command can be used to set or check a Sequell nick in the
DB for a user across discord servers. This is easier than telling the
user to go to IRC to make the mapping, although Sequell nicks are still
sometimes necessary for Discord users to perform more complicated
mappings at the crawl server level.

The !setmodrole and !removemodrole commands can be used by server admins
to set/remove a role for a given server such that users with it will
have the ability to use some commands and command arguments normal users
can't. Currently this is restricted to setting another user's
role/faction and a few joke commands, but future commands will make more
use of this.