Refactor and fix bot command code
Most of these changes are due to the changes in gammafunk/beem@9b99e15e.
Use the new 'aggregate' variable type for commands that take a role or
server as an argument, since these very often have spaces in their
names.

The various role/faction commands now have special argument names that
properly handle lookup of bot managed roles. A simple role match won't
work for these commands, since a search can match both the managed role
and the faction role corresponding to it. Also we'd like to allow the
user to specify factions either with or without ' Faction' but set the
faction role as the finalized value. Also the code now internally
refers to "managed roles" instead of "vanity roles".

Finally, the !reactbomb command has better emote lookup support so that
it can be given either the emote discord markup or just the emote name.