Fix allowed messages for direct messages
We now only allow a DM from a user if they're a member of an allowed
server. This way a user can't add the bot to an unallowed server
somewhere to just to get access to bot DM. Cache the allowance of this
in a dict so that we don't have to do the full check with every DM.

Additionally we always allow bot admins to have DM access; this is
necessary so that the bot owner can communicate with the bot before any
server has been allowed.