Add support for Discord intents
This is necessary as of 2020-10-07, per:

https://discordpy.readthedocs.io/en/latest/intents.html#intents-primer

Without the member intent, we won't be able to update the stream role,
but it seems to be affecting PMs as well. The discord.py docs don't
explicitely mention Guild.get_member() as requiring the member intent,
but it may be related to the "quirk of Discord" mentioned in the
aforementationed doc. I suspect that the member cache is not properly
updating without this intent enabled.