Add a basic output text filter
This allows a server mods to set a set of filter terms that all output
text will be checked against via a case-insensitive substring match. Any
matching instances in an output message will have the matching portions
replaced with the string "[censored]". This is very rudimentary
filtering, but will be useful enough for public discords to deal with
much of the offensive text in e.g. crawl usernames. The filter can be
checked and set via the !textfilter command and removed via the
!removetextfilter command. Users who aren't bot admins have to use these
in a guild channel.

This commit requires that the discord_servers table in the sqlite3 db be
altered accordingly:

ALTER TABLE discord_servers ADD COLUMN text_filter TEXT COLLATE NOCASE;