feat: Updates for migration to discord.py 2.0
Minor changes based on the "Migrating to v2.0" section of the migration
guide here:

https://discordpy.readthedocs.io/en/master/migrating.html

Although discord.py 2.0 is not yet release, I'm implementing and testing
the necessary changes now. The reason for this is that discord.py 1.x
uses the v7 Discord API, which is scheduled to be disabled after April
30, 2022, per:

https://github.com/discord/discord-api-docs/discussions/4510

The 2.x branch of discord.py uses the new v10 protocol, so I'm hoping to
get the necessary code changes in place and test them out before the 1.x
version of discord.py is no longer useable. The required changes were
minimal:

1) The presence updates needed for managing the Streaming role are now
received by dedicated `on_presence_update()` handler, instead of the
more general `on_member_update()` we used earlier.

2) Certain model classes no longer get in-place edits. Now the relevant
editing method returns an updated instance.

3) Receiving the contents of all guild channel messages requires a new
message content intent that must be enabled both bot-side and in code.
The idea is that verified bots will typically not use this intent,
instead using discord-native interactions like slash commands or only
get the content of messages that mention the bot, neither of which
require the message content intent. For now, since Cerebot is a relay
bot likely not aiming to get verified, we're just using message content
intent and continuing to scan all channel message. It might be good to
move the discord-native commands to slash commands and generally explore
what all we could move to discord interactions and/or usage of bot
mentions. Full details here:

https://support-dev.discord.com/hc/en-us/articles/4404772028055-Message-Content-Privileged-Intent-for-Verified-Bots

I'm putting these changes in a branch but will test live Cerebot with
this branch over the next few weeks until either the module release or
the deadline, whichever comes first.