feat: Improve error messaging
Add a `trace` argument to `log_error()`, defaulting to False, that
enables printing an exception traceback. This way we can use
`log_error()` in non-exception contexts and we don't clog up the message
log with tracebacks for simple error cases even when there is a caught
exception. Specifically some context, like login failures, are clear
enough that a simple message is sufficient. The traceback can be used
for very general contexts where we're not sure exactly what kind of
exception we'll be catching.