feat: Refactor Discord manager control logic
Previously the app used an infinite loop that handled re-instantiating
and restarting the Discord manager task in the case of any login
failure. The re-instantiation of `DiscordManager` isn't necessary nor
desired, given the workings of the discord.py `Client` class, which the
former inherits from. It's also best to have any looping/reconnect logic
contained in the manager itself to keep the app as simple as possible.

Now instantiation of `DiscordManager` is only done once, and
`DiscordManager.start()` fully controlls all aspects of login and
connection. It also has better exception handling for login errors that
differentiates between an authentication failure, which causes shut down
since we have an incorrect password, and those which are connection
issues that warrant a reconnection attempt.