feat: Refactor chat source logic
Rename a few functions to reflect that they work directly with
`DiscordSource` objects as opposed to `Channel` objects from discord.py.
Have the `DiscordManager.get_source()` method only check for a source,
returning None if not found, leaving the creation/addition of a new
source outside that function in the one place it's needed.