feat: Add support for Sequell edits
This enables the support for Sequell editing in designed channels, using
the capability added in gammafunk/beem@09cb5411. We want to enable this
on a per-channel basis, so add a channel table along with DB methods to
word with channel data. Add `!allowedits` and `!disallowedits` commands
to set the necessary permission on a per-channel basis, and use this
permission to determine if a message should be allowed to edit.