feat: Improve bot command argument handling
For discord options like 'server' and 'channel' that must ultimately
have defined values, ensure that we have these and, for channels, that
they're the right type. Move checks for definideness out of the commands
themselves and into the code that resolves these arguments in the first
place.

Also improve the lookup of channel names to allow use of channel
mentions.