feat: Improve output of the !textfilter command
If we're in a private channel, display the filters without filtering.
This is a bit silly, but the bot might get banned by another moderator
bot if it doesn't apply the filter 100% of the time in a server channel.
For private channels, we can display the unchanged filter text. There's
not a great solution for server mods wanting to check the filter;
hopefully they don't forget what their own server's filters are.