feat: Improve channel logic
Currently we don't properly detect and handle Discord threads. In
various cases, Message.channel may in fact be a thread, so we want to
make sure we work with the underlying channel of that thread. Create a
`parent_channel` property for DiscordSource that will always return this
channel and which works for non-threads as well, returning just the
usual `channel` property in that case. Use `parent_channel` where
relevant. Also properly check for threads in our `on_message` handler
for the discord client.