feat: Rewrites and updates to chat functionality
This is a major rewrite of most of Cerebot's classes due to
gammafunk/beem@d0639b03. We adopt the new `WebTilesClient` class from
the `webtiles` package as a basis for our `DiscordSource` class. The
DiscordManager and app have likewise been rewritten to follow the
read/response methodology use by `webtiles` and `beem`. There are
updates to logging and documentation as well.

We also rewrite setup.py to use a more modern `setuptools` approach and
remove the README.rst file, since setuptools can now use markdown files
directly. The README.md file has also been rewritten for clarity and to
remove outdated references. References to Freenode are replaced with
Libera in the README and the sample config file. The version variable in
version.py is now `__version__`