fix: Filter out ordered list markdown
Recently discord seems to have changed its markdown parsing to
recognized ordered lists. This messes up Sequell's listgame output,
since that often begins with a number followed by a period. The
discord.py module's `utils.escape_markdown()` function apparently
doesn't filter this even in the latest module version. I'm not sure of
what issues prevent it from doing that, but for now we'll use a regexp
substitution to escape the period with a backslash.