Check permission subsets for bot managed roles
Instead of an equality comparison of roles to those of the everyone role
for purposes of determining valid roles to manage, use a permission
subset test. The equality comparison seems to fail for unknown reasons,
whereas the subset condition is reliable and achieves what we need. A
future rework will likely move the list of bot managed roles to the DB,
but this fix works for now.